104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,14 +9,14 @@
 int main(void)
 {
 int i;
-unsigned long int a = 1, b = 2, c;
+unsigned long long int a = 1, b = 2, c;
 
-printf("%lu, %lu", a, b);
+printf("%llu, %llu", a, b);
 
 for (i = 3; i <= 98; i++)
 {
 c = a + b;
-printf(", %lu", c);
+printf(", %llu", c);
 
 a = b;
 b = c;
